# src/tsoa-auto/swagger.json
{
	"openapi": "3.0.0",
	"components": {
...
		"schemas": {
			"ResUser": {
				"properties": {
					"id": {
						"type": "number",
						"format": "double"
					},
					"name": {
						"type": "string"
					},
					"password": {
						"type": "string"
					},
					"address": {
						"type": "string"
					},
					"phone_number": {
						"type": "string"
					},
					"status": {
						"type": "boolean"
					},
					"email": {
						"type": "string"
					},
					"role": {
						"properties": {
							"description": {
								"type": "string"
							},
							"name": {
								"type": "string"
							},
							"id": {
								"type": "number",
								"format": "double"
							}
						},
						"type": "object"
					},
					"shows_count": {
						"type": "number",
						"format": "double"
					}
				},
				"type": "object",
				"additionalProperties": false
			},
...
			"ResSeasonDet": {
				"properties": {
					"id": {
						"type": "number",
						"format": "double"
					},
					"title": {
						"type": "string"
					},
					"season_no": {
						"type": "number",
						"format": "double"
					},
					"season_desc": {
						"type": "string"
					},
					"date": {
						"type": "string",
						"format": "date-time"
					},
					"episodes_count": {
						"type": "number",
						"format": "double"
					}
				},
				"type": "object",
				"additionalProperties": false
			},
			"ResShowDet": {
				"properties": {
					"id": {
						"type": "number",
						"format": "double"
					},
					"title": {
						"type": "string"
					},
					"tile_img": {
						"type": "string"
					},
					"desc_img": {
						"type": "string"
					},
					"description": {
						"type": "string"
					},
					"date": {
						"type": "string",
						"format": "date-time"
					},
					"rating": {
						"type": "string"
					},
					"show_type": {
						"type": "string"
					},
					"publisher": {
						"$ref": "#/components/schemas/ResUser"
					},
					"seasons_count": {
						"type": "number",
						"format": "double"
					},
					"seasons": {
						"items": {
							"$ref": "#/components/schemas/ResSeasonDet"
						},
						"type": "array"
					}
				},
				"required": [
					"seasons"
				],
				"type": "object",
				"additionalProperties": false
			},
			"ResShow": {
				"properties": {
					"id": {
						"type": "number",
						"format": "double"
					},
					"title": {
						"type": "string"
					},
					"tile_img": {
						"type": "string"
					},
					"desc_img": {
						"type": "string"
					},
					"description": {
						"type": "string"
					},
					"date": {
						"type": "string",
						"format": "date-time"
					},
					"rating": {
						"type": "string"
					},
					"show_type": {
						"type": "string"
					},
					"publisher": {
						"properties": {
							"email": {
								"type": "string"
							},
							"name": {
								"type": "string"
							},
							"id": {
								"type": "number",
								"format": "double"
							}
						},
						"type": "object"
					},
					"seasons_count": {
						"type": "number",
						"format": "double"
					}
				},
				"required": [
					"seasons_count"
				],
				"type": "object",
				"additionalProperties": false
			},
			"ResShowWithDetails": {
				"properties": {
					"id": {
						"type": "number",
						"format": "double"
					},
					"title": {
						"type": "string"
					},
					"tile_img": {
						"type": "string"
					},
					"desc_img": {
						"type": "string"
					},
					"description": {
						"type": "string"
					},
					"date": {
						"type": "string",
						"format": "date-time"
					},
					"rating": {
						"type": "string"
					},
					"show_type": {
						"type": "string"
					},
					"publisher": {
						"properties": {
							"email": {
								"type": "string"
							},
							"name": {
								"type": "string"
							},
							"id": {
								"type": "number",
								"format": "double"
							}
						},
						"type": "object"
					},
					"seasons_count": {
						"type": "number",
						"format": "double"
					},
					"seasons": {
						"items": {
							"properties": {
								"date": {
									"type": "string",
									"format": "date-time"
								},
								"season_desc": {
									"type": "string"
								},
								"season_no": {
									"type": "number",
									"format": "double"
								},
								"title": {
									"type": "string"
								},
								"id": {
									"type": "number",
									"format": "double"
								}
							},
							"type": "object"
						},
						"type": "array"
					}
				},
				"required": [
					"seasons_count",
					"seasons"
				],
				"type": "object",
				"additionalProperties": false
			},
			"movieRating": {
				"enum": [
					"G",
					"PG",
					"PG-13",
					"R",
					"NC-17",
					"Unrated",
					"Not Rated"
				],
				"type": "string"
			},
			"showType": {
				"enum": [
					"Movies",
					"TV Shows"
				],
				"type": "string"
			},
			"Partial_ReqShow_": {
				"properties": {
					"title": {
						"type": "string"
					},
					"tile_img": {
						"type": "string"
					},
					"desc_img": {
						"type": "string"
					},
					"description": {
						"type": "string"
					},
					"rating": {
						"$ref": "#/components/schemas/movieRating"
					},
					"show_type": {
						"$ref": "#/components/schemas/showType"
					},
					"publisher_id": {
						"type": "number",
						"format": "double"
					}
				},
				"type": "object",
				"description": "Make all properties in T optional"
			},
...
		},
		"securitySchemes": {
			"Api-Token": {
				"type": "apiKey",
				"name": "x-api-key",
				"in": "header"
			}
		}
	},
	"info": {
		"title": "Veridius-APIs",
		"version": "0.0.1",
		"description": "Veridius-APIs",
		"contact": {}
	},
	"paths": {
...
		"/shows/getAllShows": {
			"get": {
				"operationId": "GetAllShows",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"properties": {
										"totalCount": {
											"type": "number",
											"format": "double"
										},
										"items": {
											"items": {
												"$ref": "#/components/schemas/ResShowDet"
											},
											"type": "array"
										}
									},
									"required": [
										"totalCount",
										"items"
									],
									"type": "object"
								}
							}
						}
					}
				},
				"description": "Get all shows with pagination",
				"summary": "Get all shows",
				"tags": [
					"Shows"
				],
				"security": [],
				"parameters": [
					{
						"in": "query",
						"name": "page",
						"required": false,
						"schema": {
							"default": 1,
							"format": "double",
							"type": "number"
						}
					},
					{
						"in": "query",
						"name": "limit",
						"required": false,
						"schema": {
							"default": 10,
							"format": "double",
							"type": "number"
						}
					}
				]
			}
		},
		"/shows/getAllShowsFilter": {
			"get": {
				"operationId": "GetAllShowsFilter",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"properties": {
										"totalCount": {
											"type": "number",
											"format": "double"
										},
										"items": {
											"items": {
												"$ref": "#/components/schemas/ResShow"
											},
											"type": "array"
										}
									},
									"required": [
										"totalCount",
										"items"
									],
									"type": "object"
								}
							}
						}
					}
				},
				"description": "Get all shows with pagination",
				"summary": "Get all shows",
				"tags": [
					"Shows"
				],
				"security": [],
				"parameters": [
					{
						"in": "query",
						"name": "page",
						"required": false,
						"schema": {
							"default": 1,
							"format": "double",
							"type": "number"
						}
					},
					{
						"in": "query",
						"name": "limit",
						"required": false,
						"schema": {
							"default": 10,
							"format": "double",
							"type": "number"
						}
					},
					{
						"in": "query",
						"name": "showType",
						"required": false,
						"schema": {
							"type": "string"
						}
					},
					{
						"in": "query",
						"name": "rating",
						"required": false,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/shows/{showId}": {
			"get": {
				"operationId": "GetShowById",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ResShowWithDetails"
								}
							}
						}
					}
				},
				"description": "Get show by ID with details",
				"summary": "Get show details",
				"tags": [
					"Shows"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "showId",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			},
			"put": {
				"operationId": "UpdateShow",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ResShow"
								}
							}
						}
					}
				},
				"description": "Update show",
				"summary": "Update show",
				"tags": [
					"Shows"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "showId",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/Partial_ReqShow_"
							}
						}
					}
				}
			},
			"delete": {
				"operationId": "DeleteShow",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"properties": {
										"result": {
											"type": "string"
										}
									},
									"required": [
										"result"
									],
									"type": "object"
								}
							}
						}
					}
				},
				"description": "Delete show",
				"summary": "Delete show",
				"tags": [
					"Shows"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "showId",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			}
		},
		"/shows": {
			"post": {
				"operationId": "CreateShow",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ResShow"
								}
							}
						}
					}
				},
				"description": "Create a new show",
				"summary": "Create show",
				"tags": [
					"Shows"
				],
				"security": [],
				"parameters": [
					{
						"in": "query",
						"name": "userid",
						"required": false,
						"schema": {
							"format": "double",
							"type": "number"
						}
					},
					{
						"in": "query",
						"name": "title",
						"required": false,
						"schema": {
							"type": "string"
						}
					},
					{
						"in": "query",
						"name": "description",
						"required": false,
						"schema": {
							"type": "string"
						}
					},
					{
						"in": "query",
						"name": "rating",
						"required": false,
						"schema": {
							"$ref": "#/components/schemas/movieRating"
						}
					},
					{
						"in": "query",
						"name": "show_type",
						"required": false,
						"schema": {
							"$ref": "#/components/schemas/showType"
						}
					}
				],
				"requestBody": {
					"required": false,
					"content": {
						"multipart/form-data": {
							"schema": {
								"type": "object",
								"properties": {
									"tile_img": {
										"type": "string",
										"format": "binary"
									},
									"desc_img": {
										"type": "string",
										"format": "binary"
									}
								}
							}
						}
					}
				}
			}
		},
...
	},
	"servers": [
		{
			"url": "/"
		}
	],
...
}